src/github_handler.py
from github import Github
import os
from datetime import datetime
from utils import detect_framework
import git

# Define a local path for the repository
REPO_PATH = "/tmp/ai_engine_repo"

def clone_or_pull_repo():
    """
    Clones the repository if it doesn't exist locally, otherwise pulls the latest changes.
    Returns the local path to the repository.
    """
    repo_name = os.getenv("GITHUB_REPO")
    token = os.getenv("GITHUB_TOKEN")
    
    if not repo_name:
        print("[ERROR] GITHUB_REPO not found in environment variables")
        return None

    # Construct the authenticated URL
    repo_url = f"https://{token}@github.com/{repo_name}.git"

    if os.path.exists(REPO_PATH):
        try:
            print(f"[INFO] Pulling latest changes from {repo_name} into {REPO_PATH}...")
            repo = git.Repo(REPO_PATH)
            origin = repo.remotes.origin
            origin.pull()
            print("[SUCCESS] Pulled latest changes.")
        except Exception as e:
            print(f"[ERROR] Could not pull repository: {e}")
            # Fallback to re-cloning
            import shutil
            shutil.rmtree(REPO_PATH)
            return clone_or_pull_repo()
    else:
        try:
            print(f"[INFO] Cloning {repo_name} into {REPO_PATH}...")
            git.Repo.clone_from(repo_url, REPO_PATH)
            print(f"[SUCCESS] Cloned repository into {REPO_PATH}")
        except Exception as e:
            print(f"[ERROR] Could not clone repository: {e}")
            return None
            
    return REPO_PATH

from github import Github
import os
from datetime import datetime
from utils import detect_framework

def get_all_repo_files():
    """
    Recursively list all files in the repo using GitHub API.
    Ensures cross-framework support.
    """
    try:
        token = os.getenv("GITHUB_TOKEN")
        repo_name = os.getenv("GITHUB_REPO")
        
        if not token:
            print("[ERROR] GITHUB_TOKEN not found in environment variables")
            return []
        if not repo_name:
            print("[ERROR] GITHUB_REPO not found in environment variables")
            return []
            
        print(f"[DEBUG] Using GitHub token: {token[:10]}...")
        print(f"[DEBUG] Using repository: {repo_name}")
        
        g = Github(token)
        repo = g.get_repo(repo_name)
        
        def crawl(folder):
            items = []
            try:
                contents = repo.get_contents(folder)
                for c in contents:
                    if c.type == "dir":
                        items += crawl(c.path)
                    else:
                        items.append(c.path)
                return items
            except Exception as e:
                print(f"[WARNING] Error crawling folder {folder}: {e}")
                return items
                
        files = crawl("")
        print(f"[DEBUG] Found {len(files)} files in repository")
        return files
        
    except Exception as e:
        print(f"[ERROR] Failed to get repository files: {e}")
        return []

def submit_fix_pr(fixes):
    """
    Applies all code fixes (one or many) to a new branch, 
    and opens a PR with a clear description for review.
    Returns the PR URL on success, None on failure.
    """
    try:
        token = os.getenv("GITHUB_TOKEN")
        repo_name = os.getenv("GITHUB_REPO")
        
        if not token:
            print("[ERROR] GITHUB_TOKEN not found in environment variables")
            return None
        if not repo_name:
            print("[ERROR] GITHUB_REPO not found in environment variables")
            return None
            
        if not fixes:
            print("[ERROR] No fixes provided to submit_fix_pr")
            return None
            
        print(f"[DEBUG] Submitting {len(fixes)} fixes to GitHub...")
        print(f"[DEBUG] Repository: {repo_name}")
        
        g = Github(token)
        repo = g.get_repo(repo_name)
        
        # Get repository files for framework detection
        repo_files = get_all_repo_files()
        if not repo_files:
            print("[WARNING] Could not get repository files, using 'unknown' framework")
            framework = "unknown"
        else:
            framework = detect_framework(repo_files)
            print(f"[DEBUG] Detected framework: {framework}")
        
        # Get base branch
        try:
            base = repo.get_branch("main")
            print(f"[DEBUG] Using main branch as base: {base.commit.sha[:10]}...")
        except Exception as e:
            print(f"[ERROR] Could not get main branch: {e}")
            return None
        
        # Create new branch
        branch_name = f"ai-fix-{framework}-{datetime.now().strftime('%Y%m%d%H%M%S')}"
        print(f"[DEBUG] Creating branch: {branch_name}")
        
        try:
            repo.create_git_ref(ref=f"refs/heads/{branch_name}", sha=base.commit.sha)
            print(f"[SUCCESS] Created branch: {branch_name}")
        except Exception as e:
            print(f"[ERROR] Failed to create branch: {e}")
            return None
        
        # Apply fixes
        for i, fix in enumerate(fixes, 1):
            try:
                print(f"[DEBUG] Applying fix {i}/{len(fixes)}: {fix.get('path', 'unknown path')}")
                
                # Check if file exists
                try:
                    existing = repo.get_contents(fix["path"], ref=branch_name)
                    # Update existing file
                    repo.update_file(
                        path=fix["path"],
                        message=fix.get("description", f"AI fix {i}"),
                        content=fix["content"],
                        sha=existing.sha,
                        branch=branch_name
                    )
                    print(f"[SUCCESS] Updated file: {fix['path']}")
                except Exception:
                    # Create new file
                    repo.create_file(
                        path=fix["path"],
                        message=fix.get("description", f"AI fix {i}"),
                        content=fix["content"],
                        branch=branch_name
                    )
                    print(f"[SUCCESS] Created file: {fix['path']}")
                    
            except Exception as e:
                print(f"[ERROR] Failed to apply fix {i}: {e}")
                print(f"[DEBUG] Fix details: {fix}")
                continue
        
        # Create pull request with test information
        try:
            pr_title = f"AI Engine ({framework}): Automated Fixes"
            
            # Check if fixes were tested
            test_enabled = os.getenv('TEST_FIXES_BEFORE_APPLY', 'true').lower() == 'true'
            test_note = "✅ All fixes tested in isolated environment before applying" if test_enabled else "⚠️ Fix testing disabled"
            
            pr_body = f"""🔧 This PR contains {len(fixes)} code fixes automatically generated by an external AI API.

{test_note}

**Safety Features:**
- ✅ Code validation passed
- ✅ Safety checks completed
- {'✅ Isolated environment testing passed' if test_enabled else '⚠️ Testing skipped'}

**Changes:**\n"""
            for i, fix in enumerate(fixes, 1):
                pr_body += f"- {i}. {fix.get('description', 'No description')} (`{fix.get('path', 'unknown')}`)\n"
            
            pr = repo.create_pull(
                title=pr_title,
                head=branch_name,
                base="main",
                body=pr_body
            )
            
            pr_url = pr.html_url
            print(f"[SUCCESS] Pull Request Created: {pr_url}")
            return pr_url
            
        except Exception as e:
            print(f"[ERROR] Failed to create pull request: {e}")
            return None
            
    except Exception as e:
        print(f"[ERROR] submit_fix_pr failed with exception: {e}")
        import traceback
        print(f"[DEBUG] Full traceback: {traceback.format_exc()}")
        return None
