requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
PyGithub==1.59.1
google-analytics-data==0.18.0  # NEW: Google Analytics Data API
google-auth==2.23.0             # NEW: Google Authentication
google-auth-oauthlib==1.0.0     # NEW: OAuth for Google APIs
psutil==5.9.0                   # NEW: System metrics monitoring
GitPython==3.1.41
playwright==1.40.0
websockets==11.0.3
wsproto==1.2.0
